Updated cron actions to run via www-data user
Think the cron action running as root is the cause of root-owned
cache files appearing in the app, causing permission issues when
attempted to write into the same folder by the nginx/worker process.

This changes the cron setup to load the cron command for the www-data
user instead.

Related to #24